Changed the html and css - and did testing.
Here is the html, I had to change things because it looked large, weird, and stuff was shifting in mobile mode and shorter screens.
If you like it? Cool. If you hate it? You can go back to a previous version and it's all good...also I saved the previous copy just in case.
Tested everything multiple times so it should be good. 
Sorry it's so late!